# src/back-end/api/openapi/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "projects.Definition": {
        "additionalProperties": false,
        "properties": {
          "project": {
            "$ref": "#/components/schemas/domain.Project"
          },
          "roots": {
            "items": {
              "$ref": "#/components/schemas/domain.ProjectSourceRoot"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "queries.SearchQuery": {
        "additionalProperties": false,
        "properties": {
          "caseSensitive": {
            "type": "boolean"
          },
          "limit": {
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "offset": {
            "type": "integer"
          },
          "query": {
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/queries.SearchScope"
          }
        },
        "type": "object"
      },
      "queries.SearchResult": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "type": "boolean"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/search.Match"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "scannedBytes": {
            "type": "integer"
          },
          "scannedFiles": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "truncated": {
            "type": "boolean"
          },
          "viewId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "search.File": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "type": "string"
          },
          "generated": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "projectIds": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "repositoryId": {
            "type": "string"
          },
          "snapshotId": {
            "type": "string"
          },
          "symbols": {
            "items": {
              "$ref": "#/components/schemas/search.Symbol"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "test": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "search.Symbol": {
        "additionalProperties": false,
        "properties": {
          "end": {
            "type": "integer"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "start": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "valio_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "One database-bound workspace. Cookie mutations require a trusted Origin. Go type evidence is syntax-only under buildProfileId syntax-default. Latest view resolves once; pin returned viewId for subsequent requests. Maximum JSON body and scanned source are each 16 MiB; maximum 10,000 files. Project/repository IDs are explicitly registered before ingestion. Unsupported projections report unsupported. MCP is separately mounted at /mcp with bearer authentication.",
    "title": "valio.code local bootstrap API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "authentication": {
                      "type": "string"
                    },
                    "buildProfileId": {
                      "type": "string"
                    },
                    "features": {
                      "items": {
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "limits": {
                      "additionalProperties": {
                        "type": "integer"
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      }
    },
...
    "/api/v1/files/{id}": {
      "get": {
        "operationId": "getSourceFile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "viewId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/search.File"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/projects.Definition"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/projects.Definition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projects.Definition"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projects.Definition"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      },
      "put": {
        "operationId": "updateProject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/projects.Definition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projects.Definition"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      }
    },
...
    "/api/v1/search": {
      "post": {
        "operationId": "searchCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/queries.SearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/queries.SearchResult"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sanitized error; SCOPE_TOO_LARGE is never a partial success"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:8090"
    }
  ]
}